Expand docs and restructure
The introduction now contains a simple, but complete example. Also
breaks out the different sections into different articles, so
as to not overwhelm. Each usage section is expanded with a bit more
explanation of what's going on. The order of some of the sections
changed, as you don't really need to know about basic auth and retries
before you even learn how to encode/decode.