Clean up and differentiate Schema type
The `Schema` type is primarily used for a AVSC `Type` object that has
methods added for encoding/decoding the corresponding data structure.
Previous to this change, it was _also_ used to type the raw schema
that was deserialized from an `.avsc` file. This change adds a
`RawSchema` type as well as the appropriate runtime guardrails to ensure
the parsed schema is in fact valid without imposing extra restrictions
that don't apply.